requirements.txt
certifi
charset-normalizer
click
cmake
contourpy
cycler
filelock
Flask
Flask-Cors
fonttools
fsspec
idna
importlib_resources
itsdangerous
Jinja2
kiwisolver
lit
MarkupSafe
matplotlib
mpmath
networkx
numpy
opencv-python
packaging
pandas
Pillow
psutil
py-cpuinfo
pyparsing
python-dateutil
pytz
PyYAML
requests
scipy
seaborn
six
sympy
thop
torch
torchvision
tqdm
triton
typing_extensions
tzdata
ultralytics
urllib3
Werkzeug
zipp
